feat: make it work
