+++ src/commands/start.js
@@ -2,13 +2,9 @@ const { Listr } = require('listr2');
 
 const { flags: flagTypes } = require('@oclif/command');
 
-const { PrivateKey } = require('@dashevo/dashcore-lib');
-
 const BaseCommand = require('../oclif/command/BaseCommand');
 
 const MuteOneLineError = require('../oclif/errors/MuteOneLineError');
-
-const NETWORKS = require('../networks');
 
 class StartCommand extends BaseCommand {
   /**
@@ -33,10 +29,6 @@ class StartCommand extends BaseCommand {
   ) {
     const isMinerEnabled = config.get('core.miner.enable');
 
-    if (isMinerEnabled === true && config.get('network') !== NETWORKS.LOCAL) {
-      this.error(`'core.miner.interval' option supposed to work only with local network. Your network is ${config.get('network')}`, { exit: true });
-    }
-
     const tasks = new Listr(
       [
         {
@@ -48,35 +40,9 @@ class StartCommand extends BaseCommand {
               driveImageBuildPath,
               dapiImageBuildPath,
               isUpdate,
+              isMinerEnabled,
             },
           ),
-        },
-        {
-          title: 'Start a miner',
-          enabled: () => isMinerEnabled === true,
-          task: async () => {
-            let minerAddress = config.get('core.miner.address');
-
-            if (minerAddress === null) {
-              const privateKey = new PrivateKey();
-              minerAddress = privateKey.toAddress('regtest').toString();
-
-              config.set('core.miner.address', minerAddress);
-            }
-
-            const minerInterval = config.get('core.miner.interval');
-
-            await dockerCompose.execCommand(
-              config.toEnvs(),
-              'core',
-              [
-                'bash',
-                '-c',
-                `while true; do dash-cli generatetoaddress 1 ${minerAddress}; sleep ${minerInterval}; done`,
-              ],
-              ['--detach'],
-            );
-          },
         },
       ],
       {
