+++ src/listr/tasks/startNodeTaskFactory.js
@@ -1,5 +1,7 @@
 const { Listr } = require('listr2');
 
+const { PrivateKey } = require('@dashevo/dashcore-lib');
+
 const NETWORKS = require('../../networks');
 
 const wait = require('../../util/wait');
@@ -18,6 +20,7 @@ function startNodeTaskFactory(dockerCompose) {
    * @param {string} [options.driveImageBuildPath]
    * @param {string} [options.dapiImageBuildPath]
    * @param {boolean} [options.isUpdate]
+   * @param {boolean} [options.isMinerEnabled]
    * @return {Object}
    */
   function startNodeTask(
@@ -27,8 +30,13 @@ function startNodeTaskFactory(dockerCompose) {
       driveImageBuildPath = undefined,
       dapiImageBuildPath = undefined,
       isUpdate = undefined,
+      isMinerEnabled = undefined,
     },
   ) {
+    if (isMinerEnabled === true && config.get('network') !== NETWORKS.LOCAL) {
+      this.error(`'core.miner.enabled' option supposed to work only with local network. Your network is ${config.get('network')}`, { exit: true });
+    }
+
     return new Listr([
       {
         title: 'Download updated services',
@@ -64,6 +72,33 @@ function startNodeTaskFactory(dockerCompose) {
 
           // wait 10 seconds to ensure all services are running
           await wait(10000);
+        },
+      },
+      {
+        title: 'Start a miner',
+        enabled: () => isMinerEnabled === true,
+        task: async () => {
+          let minerAddress = config.get('core.miner.address');
+
+          if (minerAddress === null) {
+            const privateKey = new PrivateKey();
+            minerAddress = privateKey.toAddress('regtest').toString();
+
+            config.set('core.miner.address', minerAddress);
+          }
+
+          const minerInterval = config.get('core.miner.interval');
+
+          await dockerCompose.execCommand(
+            config.toEnvs(),
+            'core',
+            [
+              'bash',
+              '-c',
+              `while true; do dash-cli generatetoaddress 1 ${minerAddress}; sleep ${minerInterval}; done`,
+            ],
+            ['--detach'],
+          );
         },
       }]);
   }
